+++ Magick++/lib/ImageRef.cpp
@@ -92,7 +92,7 @@ void  Magick::ImageRef::options ( Magick::Options * options_ )
 }
 
 // Assign registration id to reference
-void Magick::ImageRef::id ( const long id_ )
+void Magick::ImageRef::id ( const ssize_t id_ )
 {
   if( _id > -1 )
     {
